remove unused login page, middleware testing page, combine auth types & update
