--- app/(protected)/login_renamedToPreventConflictWithAuth/page.tsx
@@ -1,7 +1,0 @@
-export default function Login() {
-  return (
-    <>
-      <h1 className="title-black">Login</h1>
-    </>
-  );
-}
