--- app/middlewareProtected/page.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-function Page() {
-  return (
-    <>
-      <div>You can only see me if you're signed in</div>
-      <div>
-        (this page & parent folder should be removed once we're done with auth
-        testing)
-      </div>
-    </>
-  );
-}
-
-export default Page;
