--- types/user/LoginSchema.ts
@@ -1,11 +1,0 @@
-import { z } from 'zod';
-
-export const loginSchema = z.object({
-  username: z
-    .string()
-    .min(3, 'Username must be at least 3 characters long.')
-    .max(20, 'Username must be less than 20 characters long.'),
-  password: z.string().min(8, 'Password must be atleast 8 characters long.'),
-});
-
-export type LoginSchemaType = z.infer<typeof loginSchema>;
