--- types/user/SignUpResponse.ts
@@ -1,3 +1,0 @@
-type SignUpResponse = {
-  token: string;
-};
